pyproject.toml
[project]
name = "sector-leadlag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
